Work around a few Visual C++ 2017 issues.

+++ selene/img/pixel/PixelTraits.hpp
@@ -27,7 +27,7 @@ template <typename Element_>
 struct PixelTraits
 {
   using Element = Element_;  ///< The pixel element type.
-  static constexpr std::ptrdiff_t nr_channels = 1;  ///< The number of channels per pixel.
+  static constexpr std::int16_t nr_channels = 1;  ///< The number of channels per pixel.
   static constexpr std::int16_t nr_bytes = sizeof(Element);  ///< The number of bytes per pixel.
   static constexpr std::int16_t nr_bytes_per_channel = sizeof(Element);  ///< The number of bytes per pixel channel.
 
@@ -63,7 +63,7 @@ template <typename T, std::size_t N, PixelFormat pixel_format_>
 struct PixelTraits<Pixel<T, N, pixel_format_>>
 {
   using Element = T;  ///< The pixel element type.
-  static constexpr std::ptrdiff_t nr_channels = N;  ///< The number of channels per pixel.
+  static constexpr std::int16_t nr_channels = N;  ///< The number of channels per pixel.
   static constexpr std::int16_t nr_bytes = sizeof(Pixel<T, N, pixel_format_>);  ///< The number of bytes per pixel.
   static constexpr std::int16_t nr_bytes_per_channel = sizeof(Element);  ///< The number of bytes per pixel channel.
 
@@ -101,7 +101,7 @@ template <typename T, std::size_t N, PixelFormat pixel_format_>
 struct PixelTraits<const Pixel<T, N, pixel_format_>>
 {
   using Element = T;  ///< The pixel element type.
-  static constexpr std::size_t nr_channels = N;  ///< The number of channels per pixel.
+  static constexpr std::int16_t nr_channels = N;  ///< The number of channels per pixel.
   static constexpr std::int16_t nr_bytes = sizeof(const Pixel<T, N, pixel_format_>);  ///< The number of bytes per pixel.
   static constexpr std::int16_t nr_bytes_per_channel = sizeof(Element);  ///< The number of bytes per pixel channel.
